requirements.txt
Django==1.11.16
Pillow==5.3.0
psycopg2-binary==2.7.5
pytz==2018.7
